+++ marketing_event-20210101/alibabacloud_marketing_event20210101/models/_query_qwen_conference_sg_ticket_pop_response_body.py
@@ -14,7 +14,6 @@ class QueryQwenConferenceSgTicketPopResponseBody(DaraModel):
         request_id: str = None,
         success: bool = None,
     ):
-        # Id of the request
         self.code = code
         self.data = data
         self.message = message