+++ marketing_event-20210101/alibabacloud_marketing_event20210101/models/_query_session_list_pop_response_body.py
@@ -100,14 +100,10 @@ class QuerySessionListPopResponseBodyData(DaraModel):
         name: str = None,
         start_time: str = None,
     ):
-        # code
         self.code = code
         self.end_time = end_time
-        # id
         self.id = id
-        # location
         self.location = location
-        # name
         self.name = name
         self.start_time = start_time