+++ marketing_event-20210101/alibabacloud_marketing_event20210101/models/_query_sign_in_record_pop_response_body.py
@@ -99,9 +99,7 @@ class QuerySignInRecordPopResponseBodyData(DaraModel):
         time: str = None,
     ):
         self.event = event
-        # nfcid
         self.rfid = rfid
-        # sessionId
         self.session_id = session_id
         self.time = time